Fix hashtable init bug - can init bigger now!

+++ src/pvtable.h
@@ -11,7 +11,7 @@
 enum { HFNONE, HFALPHA, HFBETA, HFEXACT };
 
 
-void InitHashTable(S_HASHTABLE *table, const int MB);
+void InitHashTable(S_HASHTABLE *table, const uint64_t MB);
 void StoreHashEntry(S_BOARD *pos, const int move, int score, const int flags, const int depth);
 int ProbeHashEntry(S_BOARD *pos, int *move, int *score, int alpha, int beta, int depth);
 int GetPvLine(const int depth, S_BOARD *pos);
